1.我的资源;2.查询已授权从帐号;3.创建单点会话;4.解决死锁;

+++ server/biz/model/dto/sys_menu.go
@@ -7,6 +7,6 @@ type SysMenuSaveReq struct {
 	Path       string `json:"path" binding:"max=64"`
 	MenuType   int8   `json:"menuType" binding:"oneof=0 1 2"`
 	Permission string `json:"permission" binding:"required,max=64"`
-	Hidden     int8   `json:"hidden"`
+	Hidden     int16  `json:"hidden"`
 	Component  string `json:"component"`
 }
